Treat angle values as floating point numbers
This avoids having the floor suddenly tilt when the angle changes from 0.9 to 1.

From: Alberto Mardegan <mardy@users.sourceforge.net>

git-svn-id: https://s.snth.net/svn/neverball/trunk@3612 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ ball/game_server.c
@@ -823,7 +823,7 @@ void game_set_z(float k)
     input_set_z(+ANGLE_BOUND * k);
 }
 
-void game_set_ang(int x, int z)
+void game_set_ang(float x, float z)
 {
     input_set_x(x);
     input_set_z(z);
